chore(transform): move file to compat directory
